sync: fix PTP config and run sync (#1214)
A recent commit to the webhooks made some structs exported, so that they
show up in the generated deepcopy file. However, the PTP sync config
excluded the webhook files. This commit adds them back in so the
compilation errors are resolved. Fortunately, there are no issues with
the webhook and our controller-runtime version.